poke-agent/requirements.txt
python-dotenv
google-cloud-aiplatform
google-generativeai
langgraph
langchain-google-genai
langchain_core
rich
poke-env
